refactor(service): 拆分内容治理导航装配逻辑
