+++ Docs/planning/current.md
@@ -8,7 +8,7 @@
 
 - **阶段**：`第三开发阶段：真实使用增长与长期契约治理`
 - **当前主线**：`P3-7-C3 后端 Service 热区评估与首批治理候选`
-- **复核日期**：`2026-05-22`
+- **复核日期**：`2026-05-23`
 - **最近结论**：
   - `P3-1` 至 `P3-5` 已完成公开内容增长、PublicId 试点、留存回流、动态 sitemap 与详情 head snapshot 首批闭环
   - `P3-6` 公开增长部署观察已收口，本地 Gateway 与生产公开域名 `https://radishx.com` 的 public head smoke 均通过，转入维护线
@@ -17,6 +17,7 @@
   - `P3-7-C` 已从原规划、backlog、未完成功能和近期主路径中选出下一批一天级任务：先做 WikiApp 热区治理，再做 ChatApp 热区治理
   - `P3-7-C1` 已完成 WikiApp 文档工作台首批热区拆分：侧边栏抽出为独立组件，纯 helper 迁出主文件，`WikiApp.tsx` 已降至硬上限以内
   - `P3-7-C2` 已完成 ChatApp 聊天工作台首批热区拆分：消息列表、频道侧栏、成员面板、输入区状态和纯 helper 已从主文件抽出，`ChatApp.tsx` 已降至硬上限以内
+  - `P3-7-C3` 已完成后端 Service 首批候选治理：`ContentModerationService` 采用行为等价 partial 分文件拆出举报目标快照解析与队列导航装配，主文件从 `1735` 行降至 `806` 行
   - 项目仍处于单人开发期和功能建设期，没有稳定用户反馈和专职测试，不能把“等待真实使用观察”作为默认主线
 
 ## 当前执行入口
@@ -33,7 +34,7 @@
 
 1. **收口工作台代码热区首批治理**
    - `WikiApp -> ChatApp` 两个一天级前端热区任务已完成首批行为等价拆分
-   - 下一步优先回到第三阶段已点名的后端 Service 热区，先做窄范围评估，再决定是否启动一天级拆分
+   - 后端 Service 首批已完成 `ContentModerationService` 行为等价拆分，下一步继续复核 `ExperienceService.cs` 是否存在一天级、安全可验证的拆分点
    - 不把“观察”作为无事可做的结论，也不直接把后续重点提前升级为当前主线
 2. **维护观察降级为并行线**
    - `P3-6 / P3-7` 暴露的新问题仍可回拉小闭环，但不再占用主线
@@ -46,7 +47,7 @@
 ## 下一顺位
 
 - `P3-7-C3 后端 Service 热区评估与首批治理候选`
-  - 优先复核 `ExperienceService.cs`、`ContentModerationService.cs` 等第三阶段已点名热区，选出一个一天级、行为等价、可验证的拆分候选
+  - 已完成 `ContentModerationService.cs` 首批行为等价拆分；下一步优先复核 `ExperienceService.cs` 治理规则、快照映射与纯 helper 是否存在安全拆分候选
   - 不改 API 契约、权限语义、数据库结构或业务规则；若评估发现需要运行时行为变更，应先单独评审
   - 验证优先覆盖对应后端定向测试与 `dotnet build Radish.slnx -c Debug`
 - `P3-8 多端功能补全与 UI 设计治理` 保留为后续重点方向，不作为当前最近任务的默认入口
